Implement named context dependent dot-handling
This was a little tricky but it works now, afaik. Whenever we're in an
array and we discover a { or [, we need to make sure that the name,
if empty, is not printed, because that messes up the dot identifiers.

I see that I need to add additional cehcks for $2 to be empty.